python-worker/requirements.txt
videosdk==0.0.5
pillow==10.4.0
python-dotenv==1.0.1
asyncio==3.4.3
